Added several new subcommands of /general - restrict <node> - adds a permissions node to the ops-only list - release <node> - removes a permissions node from the ops-only list - kit <kit> add [<item> [<amount>]] - adds items to a kit, creating it if it doesn't exist - kit <kit> remove <item> [<amount>] - removes items from a kit - kit <kit> delay <delay> - sets the delay of a kit - kit <kit> cost <cost> - sets the cost of a kit - kit <kit> trash - completely deletes a kit - kit <kit> list - lists the contents of a kit - economy <node> <cost> - not yet implemented; it'll be used to set the cost of an economy node - set <var> <value> - not yet implemented; it'll be used to set some of the simpler config options - item group <group> add|remove <item> - not yet implemented; it'll be used to alter the contents of the /give white/blacklists Other changes - Added comments to the command-helper.txt - Added "auto-save" option to the config; if true, kits.yml and config.yml are saved when the server is stopped; items.yml is always saved regardless of this setting. - Each /general subcommand now has its own permissions node. - Removed some unnecessary casts - Began deprecating stuff that will be removed in 4.0

+++ src/net/craftstars/general/command/info/whoCommand.java
@@ -47,7 +47,7 @@ public class whoCommand extends CommandBase {
 	
 	private String getHealthBar(Player ofWhom) {
 		String healthBar = "[";
-		double health = ((double) ofWhom.getHealth()) / 2.0;
+		double health = ofWhom.getHealth() / 2.0;
 		String colour = health < 3 ? "&c" : (health < 7 ? "&e" : "&2");
 		healthBar += colour;
 		healthBar += Toolbox.repeat('|', (int) health);
